requirements.txt
channels==4.1.0
Django==5.0.6
portkey-ai==1.3.2
python-dotenv==1.0.1
psycopg2==2.9.9
daphne==4.1.2
django-recaptcha==4.0.0
google-api-python-client==2.132.0